result table and download

+++ wenwan-data-service/src/main/java/com/wenwan/service/api/parse/TaskService.java
@@ -21,7 +21,7 @@ public interface TaskService {
 
     void insertSql(TaskSqlVo taskSqlVo);
 
-    int updateSql(TaskSqlVo taskSqlVo);
+    Long updateSql(TaskSqlVo taskSqlVo);
 
     int deleteSql(Long sqlId);
